requirements.txt
fastapi
uvicorn
python-jose[cryptography]
PyJWT
requests
sqlmodel
